+++ src/Test/Integration/TestBase.cs
@@ -1,21 +1,3 @@
-// <copyright file="TestBase.cs" company="Basho Technologies, Inc.">
-// Copyright 2015 - Basho Technologies, Inc.
-//
-// This file is provided to you under the Apache License,
-// Version 2.0 (the "License"); you may not use this file
-// except in compliance with the License.  You may obtain
-// a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-// </copyright>
-
 namespace Test.Integration
 {
     using System;
@@ -46,12 +28,12 @@ namespace Test.Integration
             RiakClient.DisableListBucketsWarning = true;
         }
 
-        public TestBase()
+        public TestBase(bool auth = true)
         {
 #if NOAUTH
             cluster = RiakCluster.FromConfig("riak1NodeNoAuthConfiguration");
 #else
-            if (MonoUtil.IsRunningOnMono)
+            if (auth == false || MonoUtil.IsRunningOnMono)
             {
                 cluster = RiakCluster.FromConfig("riak1NodeNoAuthConfiguration");
             }
@@ -59,32 +41,36 @@ namespace Test.Integration
             {
                 cluster = RiakCluster.FromConfig("riak1NodeConfiguration");
             }
+
 #endif
             client = cluster.CreateClient();
         }
 
-        protected abstract RiakString BucketType
+        protected virtual RiakString BucketType
         {
-            get;
+            get { return "bucketType_unset"; }
         }
 
-        protected abstract RiakString Bucket
+        protected virtual RiakString Bucket
         {
-            get;
+            get { return "bucket_unset"; }
         }
 
         [TestFixtureSetUp]
-        protected virtual void TestFixtureSetUp()
+        public virtual void TestFixtureSetUp()
         {
         }
 
         [TestFixtureTearDown]
-        protected virtual void TestFixtureTearDown()
+        public virtual void TestFixtureTearDown()
         {
-            foreach (string key in Keys)
+            if (EnumerableUtil.NotNullOrEmpty(Keys))
             {
-                var id = new RiakObjectId(BucketType, Bucket, key);
-                client.Delete(id);
+                foreach (string key in Keys)
+                {
+                    var id = new RiakObjectId(BucketType, Bucket, key);
+                    client.Delete(id);
+                }
             }
         }
